refactor: consolidate chain configs
