+++ apps/maestro/src/lib/utils/getNativeToken.ts
@@ -1,7 +1,6 @@
 import { memoize } from "@axelarjs/utils";
 
-import { WAGMI_CHAIN_CONFIGS } from "~/config/evm-chains";
-import { VM_CHAIN_CONFIGS } from "~/config/vm-chains";
+import { CHAIN_CONFIGS } from "~/config/chains";
 
 /**
  * Overrides for native tokens that have a symbol mismatch with the gas estimation API
@@ -24,7 +23,7 @@ export const getNativeToken = memoize((axelarChainId: string) => {
     return NATIVE_TOKEN_OVERRIDES[axelarChainId];
   }
 
-  const chainConfig = [...WAGMI_CHAIN_CONFIGS, ...VM_CHAIN_CONFIGS].find(
+  const chainConfig = CHAIN_CONFIGS.find(
     (chain) => chain.axelarChainId.toLowerCase() === axelarChainId.toLowerCase()
   );
